requirements.txt
flask
openai
python-dotenv
scikit-learn
